clarify confusing formulation
the paragraph is confusing for newcomers for the following reasons:

it sais calling fetch is an antipattern, but it would just not work. in addtion the solution to use service binding also calls fetch, just not the global fetch function but the fetch function of the service binding so this will confuse users even more. lastly, "referencing external dependencies with a custom response" sounds like a special cloudflare feature, when all that is meant is redirecting to another url which is super hard to understand